Create class ProdutoController
