+++ src/main/java/br/com/fiap/soat/techChallenge/api/PedidoApi.java
@@ -3,9 +3,6 @@ package br.com.fiap.soat.techChallenge.api;
 import br.com.fiap.soat.techChallenge.api.requests.PedidoRequest;
 import br.com.fiap.soat.techChallenge.controllers.PedidoController;
 import br.com.fiap.soat.techChallenge.responses.PedidoResponse;
-import br.com.fiap.soat.techChallenge.entities.Pedido;
-import br.com.fiap.soat.techChallenge.interfaces.usecases.FazerPedidoUseCasePort;
-import br.com.fiap.soat.techChallenge.interfaces.usecases.ObterPedidosUseCasePort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
