+++ src/main/java/br/com/fiap/soat/techChallenge/config/UseCaseBeanConfig.java
@@ -1,29 +1,15 @@
 package br.com.fiap.soat.techChallenge.config;
 
-import br.com.fiap.soat.techChallenge.controllers.ClienteController;
-import br.com.fiap.soat.techChallenge.controllers.PedidoController;
-import br.com.fiap.soat.techChallenge.interfaces.usecases.*;
 import br.com.fiap.soat.techChallenge.interfaces.gateways.ClienteGatewayPort;
-import br.com.fiap.soat.techChallenge.interfaces.gateways.ProdutoGatewayPort;
 import br.com.fiap.soat.techChallenge.interfaces.gateways.PedidoGatewayPort;
-
+import br.com.fiap.soat.techChallenge.interfaces.gateways.ProdutoGatewayPort;
+import br.com.fiap.soat.techChallenge.interfaces.usecases.*;
 import br.com.fiap.soat.techChallenge.usecases.*;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class BeanConfig {
-
-    @Bean
-    public ClienteController clienteController(CadastrarClienteUseCasePort cadastrarClienteUseCase,
-                                               IdentificarClienteUseCasePort identificarClienteUseCase) {
-        return new ClienteController(cadastrarClienteUseCase, identificarClienteUseCase);
-    }
-
-    @Bean
-    public PedidoController pedidoController(FazerPedidoUseCasePort fazerPedidoUseCase, ObterPedidosUseCasePort obterPedidosUseCase) {
-        return new PedidoController(fazerPedidoUseCase, obterPedidosUseCase);
-    }
+public class UseCaseBeanConfig {
 
     @Bean
     public CadastrarClienteUseCasePort cadastrarClienteUseCasePort(ClienteGatewayPort clienteGatewayPort) {
